requirements.txt
Flask==2.3.2
spotipy==2.23.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
